refactor(ds): merge button components and story
