--- ui/design-system-core/src/components/Button/index.ts
@@ -1,2 +1,0 @@
-export { Button } from './Button';
-export { ButtonAlt } from './ButtonAlt';
